requirements.txt
anthropic
google-generativeai
openai
groq
pypdf
